email service commit

+++ src/main/java/com/amitit/webapp/entity/User.java
@@ -1,5 +1,6 @@
 package com.amitit.webapp.entity;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import jakarta.persistence.Column;
@@ -19,24 +20,24 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Table(name = "users")
 public class User {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long uid;
-	
+	private int uid;
+
 	private String name;
 	private String email;
 	private String password;
 //	private String address;
 	private String contact;
 	// This is sample comment
-	
+
 //	@Column(name = "aadhaar_no")
 	private String aadhaarNo;
-	
+
 //	@Column(name = "photo_id")
 	private String photoId;
-	
+
 	@OneToMany(mappedBy = "user")
 	private List<Enrollment> enrollments;
